JFYI: mark "links" directory as a sources root in IDEA.
As you see, pretty hacky and not obvious. And it refers to "java", that is not semantically correct.

The other way to mark it as sources root would be to apply `idea` plugin and modify module's XML (`awesome-kotlin.iml`), using one of lifecycle hooks from https://docs.gradle.org/current/userguide/idea_plugin.html