chore: Update Gradle distribution URL and SHA256 sum
Updated the Gradle distribution URL in `gradle-wrapper.properties` to `https://services.gradle.org/distributions/gradle-8.4-bin.zip`. Also updated the SHA256 sum for the Gradle distribution. Additionally, made a small change in `gradlew` to discard the standard output of the `cd` command to `/dev/null` in order to prevent printing the working directory when `$CDPATH` is set.